public modifier removed for better readability (junit 5) unit test update for change password and delete user

+++ twitter/src/main/java/ro/info/iasi/fiipractic/twitter/service/UserService.java
@@ -2,24 +2,31 @@ package ro.info.iasi.fiipractic.twitter.service;
 
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
+import ro.info.iasi.fiipractic.twitter.exception.BadRequestException;
 import ro.info.iasi.fiipractic.twitter.exception.NotFoundException;
 import ro.info.iasi.fiipractic.twitter.exception.UsernameTakenException;
 import ro.info.iasi.fiipractic.twitter.model.User;
 import ro.info.iasi.fiipractic.twitter.repository.UserJpaRepository;
+import ro.info.iasi.fiipractic.twitter.util.IPasswordEncoder;
+import ro.info.iasi.fiipractic.twitter.util.PasswordEncoder;
 
 import java.util.List;
 
 @Service
 public class UserService {
     private final UserJpaRepository userRepository;
-    public UserService(UserJpaRepository userRepository) {
+
+    private final IPasswordEncoder passwordEncoder;
+    public UserService(UserJpaRepository userRepository, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
+        this.passwordEncoder = passwordEncoder;
     }
 
 
     public User saveUser(User user){
         if(userRepository.findByUsername(user.getUsername())!=null)
             throw new UsernameTakenException("Username " + user.getUsername() + " is taken.");
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         return userRepository.save(user);
     }
 
@@ -27,6 +34,7 @@ public class UserService {
         User user = userRepository.findByUsername(username);
         if(user == null)
             throw new NotFoundException("User with username '" + username + "' was not found.");
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         return user;
     }
 
@@ -37,16 +45,26 @@ public class UserService {
             throw new NotFoundException("No users found with the specified search criteria.");
         return foundUsers;
     }
-
-    public void deleteUser(User user) {
+    public void deleteUser(User user, String password) {
+        if(!passwordEncoder.matches(password, user.getPassword()))
+            throw new BadRequestException("Password provided is incorrect.");
         userRepository.delete(user);
     }
     @Transactional
     public void updateUser(User user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         userRepository.updateUser(user.getId(),
                 user.getFirstname(),
                 user.getLastname(),
                 user.getEmail(),
                 user.getPassword());
+    }
+
+    @Transactional
+        public void updateUserPassword(User user, String oldPassword, String newPassword) {
+        if(!passwordEncoder.matches(oldPassword, user.getPassword()))
+            throw new BadRequestException("Old password is incorrect.");
+        user.setPassword(passwordEncoder.encode(newPassword));
+        userRepository.updateUserPassword(user.getId(), user.getPassword());
     }
 }
